### 2020-05-22 #2230 @whedon
[PRE REVIEW]: dcgp: Differentiable Cartesian Genetic Programming made easy.
**Submitting author:** @darioizzo (<a href="http://orcid.org/0000-0002-9846-8423">Dario Izzo</a>)
**Repository:** <a href="https://github.com/darioizzo/dcgp" target ="_blank">https://github.com/darioizzo/dcgp</a>
**Version:** v1.5.1
