### 2020-05-23 #2230 @darioizzo
@arfon I would rather keep the repositories for the code and this paper separate. The reasons are that the code will evolve and change, while the paper, if published should stay unmodified. I also do not like having too many files in the root folder of the code repository (but thats a matter of taste).
