### 2020-05-24 #2230 @arfon
> I could make a branch with the paper only if thats better?

Yes, a branch would be better than what we have currently here. Also, the paper can be anywhere in the repository (i.e. nested in a folder) so doesn't need to be in the repository root.
