+++ include/dcgp/wrapped_functions.hpp
@@ -1,632 +1,0 @@
-#ifndef DCGP_WRAPPED_FUNCTIONS_H
-#define DCGP_WRAPPED_FUNCTIONS_H
-
-#include <audi/audi.hpp>
-#include <audi/functions.hpp>
-#include <cmath>
-#include <string>
-#include <type_traits>
-#include <vector>
-
-#include <dcgp/config.hpp>
-#include <dcgp/function.hpp>
-#include <dcgp/type_traits.hpp>
-
-#define DCGP_S11N_FUNCTION_EXPORT_KEY_MULTI(f)                                                                         \
-    DCGP_S11N_FUNCTION_EXPORT_KEY(dcgp_##f##_double, dcgp::f##_func<double>, double, const std::vector<double> &)      \
-    DCGP_S11N_FUNCTION_EXPORT_KEY(dcgp_##f##_gdual_d, dcgp::f##_func<audi::gdual_d>, audi::gdual_d,                    \
-                                  const std::vector<audi::gdual_d> &)                                                  \
-    DCGP_S11N_FUNCTION_EXPORT_KEY(dcgp_##f##_gdual_v, dcgp::f##_func<audi::gdual_v>, audi::gdual_v,                    \
-                                  const std::vector<audi::gdual_v> &)
-
-#define DCGP_S11N_FUNCTION_IMPLEMENT_MULTI(f)                                                                          \
-    DCGP_S11N_FUNCTION_IMPLEMENT(dcgp_##f##_double, dcgp::f##_func<double>, double, const std::vector<double> &)       \
-    DCGP_S11N_FUNCTION_IMPLEMENT(dcgp_##f##_gdual_d, dcgp::f##_func<audi::gdual_d>, audi::gdual_d,                     \
-                                 const std::vector<audi::gdual_d> &)                                                   \
-    DCGP_S11N_FUNCTION_IMPLEMENT(dcgp_##f##_gdual_v, dcgp::f##_func<audi::gdual_v>, audi::gdual_v,                     \
-                                 const std::vector<audi::gdual_v> &)
-
-#define DCGP_S11N_FUNCTION_EXPORT_KEY_STRING(f)                                                                        \
-    DCGP_S11N_FUNCTION_EXPORT_KEY(dcgp_##f##_string, dcgp::f##_func, std::string, const std::vector<std::string> &)
-
-#define DCGP_S11N_FUNCTION_IMPLEMENT_STRING(f)                                                                         \
-    DCGP_S11N_FUNCTION_IMPLEMENT(dcgp_##f##_string, dcgp::f##_func, std::string, const std::vector<std::string> &)
-
-#define DCGP_S11N_EMPTY_SERIALIZE_MEMFN()                                                                              \
-    template <typename Archive>                                                                                        \
-    void serialize(Archive &, unsigned)                                                                                \
-    {                                                                                                                  \
-    }
-
-namespace dcgp
-{
-
-// SFINAE dust (to hide under the carpet). Its used to enable the templated
-// version of the various functions that can construct a kernel object. Only for
-// double and a gdual type Complex could also be allowed.
-template <typename T>
-using f_enabler = typename std::enable_if<std::is_same<T, double>::value || is_gdual<T>::value, int>::type;
-
-/*--------------------------------------------------------------------------
[…]
