+++ dcgpy/test.py
@@ -1,555 +1,0 @@
-from __future__ import absolute_import as _ai
-
-import unittest as _ut
-
-
-def my_py_fun(x):
-    return sum(x)
-
-
-def my_py_fun_print(x):
-    s = "+"
-    return "(" + s.join(x) + ")"
-
-
-class test_kernel(_ut.TestCase):
-    def runTest(self):
-        self.test_double()
-        self.test_gdual_double()
-        self.test_gdual_vdouble()
-
-    def test_double(self):
-        from dcgpy import kernel_double as kernel
-
-        my_kernel = kernel(my_py_fun, my_py_fun_print, "my_sum_kernel")
-
-        self.assertEqual(my_kernel.__repr__(), "my_sum_kernel")
-        self.assertEqual(my_kernel([1, 2, 3]), 6)
-        self.assertEqual(my_kernel(["x", "y"]), "(x+y)")
-
-    def test_gdual_double(self):
-        from dcgpy import kernel_gdual_double as kernel
-        from pyaudi import gdual_double as gdual
-
-        my_kernel = kernel(my_py_fun, my_py_fun_print, "my_sum_kernel")
-
-        self.assertEqual(my_kernel.__repr__(), "my_sum_kernel")
-        x = gdual(1, "x", 2)
-        y = gdual(2, "y", 2)
-        z = gdual(3, "z", 2)
-        self.assertEqual(my_kernel([x, y, z]), x + y + z)
-        self.assertEqual(my_kernel([x, y, z]), my_kernel([x, y, z]))
-        self.assertEqual(my_kernel(["x", "y"]), "(x+y)")
-
-    def test_gdual_vdouble(self):
-        from dcgpy import kernel_gdual_vdouble as kernel
-        from pyaudi import gdual_vdouble as gdual
-
-        my_kernel = kernel(my_py_fun, my_py_fun_print, "my_sum_kernel")
-
-        self.assertEqual(my_kernel.__repr__(), "my_sum_kernel")
-        x = gdual([1, -1], "x", 2)
-        y = gdual([2, -2], "y", 2)
-        z = gdual([-2, 1], "z", 2)
-        self.assertEqual(my_kernel([x, y, z]), x + y + z)
-        self.assertEqual(my_kernel([x, y, z]), my_kernel([x, y, z]))
-        self.assertEqual(my_kernel(["x", "y"]), "(x+y)")
-
-    def test_serialization_double(self):
-        import cloudpickle as cpk
-        from dcgpy import kernel_set_double as kernel_set
-        from dcgpy import kernel_double as kernel
-
-        # cpp kernels
-        cpp_kv = kernel_set(
-            ["sum", "diff", "mul", "div", "tanh", "sig", "cos", "sin", "log", "exp", "gaussian", "sqrt", "ReLu", "ELU", "ISRU"])()
-        x1 = 1.2
-        x2 = -1.2
-        x3 = 3.2
-
-        for cpp_k in cpp_kv:
-            cpp_k2 = cpk.loads(cpk.dumps(cpp_k))
-            self.assertEqual(cpp_k([x1, x2, x3]), cpp_k2([x1, x2, x3]))
-            self.assertEqual(cpp_k(["a", "b", "c"]), cpp_k2(["a", "b", "c"]))
-
-        # pythonic kernels
-        my_py_kernel = kernel(my_py_fun, my_py_fun_print, "my_py_fun")
-        my_py_kernel2 = cpk.loads(cpk.dumps(my_py_kernel))
-        self.assertEqual(my_py_kernel(
-            [x1, x2, x3]), my_py_kernel([x1, x2, x3]))
-        self.assertEqual(my_py_kernel(
-            ["a", "b", "c"]), my_py_kernel2(["a", "b", "c"]))
-
[…]
