+++ paper.bib
@@ -1,0 +1,56 @@
+@inproceedings{izzo:2017,
+  title={Differentiable genetic programming},
+  author={Izzo, Dario and Biscani, Francesco and Mereta, Alessio},
+  booktitle={European Conference on Genetic Programming},
+  pages={35--51},
+  year={2017},
+  organization={Springer}
+}
+
+@book{schwefel:1993,
+  title={Evolution and optimum seeking: the sixth generation},
+  author={Schwefel, Hans-Paul Paul},
+  year={1993},
+  publisher={John Wiley \& Sons, Inc.}
+}
+
+@article{koza:2010,
+  title={Human-competitive results produced by genetic programming},
+  author={Koza, John R},
+  journal={Genetic programming and evolvable machines},
+  volume={11},
+  number={3-4},
+  pages={251--284},
+  year={2010},
+  publisher={Springer}
+}
+
+@inproceedings{miller:2011,
+  title={Cartesian genetic programming},
+  author={Miller, Julian F},
+  booktitle={Cartesian Genetic Programming},
+  pages={17--34},
+  year={2011},
+  organization={Springer}
+}
+
+@misc{audi:2020,
+  author       = {Dario Izzo and
+                  Francesco Biscani},
+  title        = {AuDi: v1.8},
+  year         = 2020,
+  journal      = {Zenodo},
+  version      = {v1.8},
+  doi          = {10.5281/zenodo.3702783},
+  url          = {https://doi.org/10.5281/zenodo.3702783}
+}
+
+@inproceedings{martens:2019,
+  title={Neural network architecture search with differentiable cartesian genetic programming for regression},
+  author={M{\"a}rtens, Marcus and Izzo, Dario},
+  booktitle={Proceedings of the Genetic and Evolutionary Computation Conference Companion},
+  pages={181--182},
+  year={2019},
+  organization={Association for Computing Machinery (ACM)}
+}
+
