+++ README.md
@@ -1,0 +1,3 @@
+# dcgp-paper
+
+This repository contains the necessary files for submitting the paper describing dcgpy to the Journal of Open Source Software JOSS.
