### 2020-05-26 #2230 @darioizzo
@arfon I have moved all files to a branch in the repository https://github.com/darioizzo/dcgp/tree/joss-paper-v1.5.1
