### 2020-05-26 #2230 @darioizzo
proof looks ok to me.
