### 2020-05-28 #2230 @danielskatz
👋 @VivianePons - would you be able to edit this submission for JOSS?
