### 2020-06-05 #2290 @whedon
[REVIEW]: dcgp: Differentiable Cartesian Genetic Programming made easy.
**Submitting author:** @darioizzo (<a href="http://orcid.org/0000-0002-9846-8423">Dario Izzo</a>)
**Repository:** <a href="https://github.com/darioizzo/dcgp" target ="_blank">https://github.com/darioizzo/dcgp</a>
**Version:** 1.5
