### 2020-06-05 #2290 @VivianePons
For info, the paper is in a branch here: https://github.com/darioizzo/dcgp/tree/joss-paper-v1.5.1
