### 2020-06-06 #2290 @shah314
For the authors:

Can you add the missing DOIs to the paper?

Can you add a brief description of audi to the paper (just to make the paper complete and readable), since audi looks like a codebase on Zenodo.

Do you do any form of crossover between A and B? Like replacing a random subtree of A with a random subtree of B? I could not find anything in the paper.

Any results, or comparison with benchmarks or do you think it is not necessary? For instance, you could compare the results with linear or quadratic regression?
