+++ paper.bib
@@ -4,14 +4,16 @@
   booktitle={European Conference on Genetic Programming},
   pages={35--51},
   year={2017},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/978-3-319-55696-3_3}
 }
 
 @book{schwefel:1993,
   title={Evolution and optimum seeking: the sixth generation},
   author={Schwefel, Hans-Paul Paul},
   year={1993},
-  publisher={John Wiley \& Sons, Inc.}
+  publisher={John Wiley \& Sons, Inc.},
+  isbn={978-0-471-57148-3}
 }
 
 @article{koza:2010,
@@ -22,7 +24,8 @@
   number={3-4},
   pages={251--284},
   year={2010},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s10710-010-9112-3}
 }
 
 @inproceedings{miller:2011,
@@ -42,7 +45,6 @@
   journal      = {Zenodo},
   version      = {v1.8},
   doi          = {10.5281/zenodo.3702783},
-  url          = {https://doi.org/10.5281/zenodo.3702783}
 }
 
 @inproceedings{martens:2019,
@@ -51,6 +53,7 @@
   booktitle={Proceedings of the Genetic and Evolutionary Computation Conference Companion},
   pages={181--182},
   year={2019},
-  organization={Association for Computing Machinery (ACM)}
+  organization={Association for Computing Machinery (ACM)},
+  doi={10.1145/3319619.3322003}
 }
 
