### 2020-06-08 #2290 @Ohjeah
#### Software paper

- The introduction is a little bit steep for non-experts. As an outsider I would not be able to understand it (e.g. 1. sentence: What is a fitness?). Could you please add a paragraph for non-experts?
- I am missing the statement of need as well as the state of the field. I think you can repurpose parts of the "Comparison to the CGP-Library" paragraph from the README.
