+++ paper.bib
@@ -34,7 +34,8 @@
   booktitle={Cartesian Genetic Programming},
   pages={17--34},
   year={2011},
-  organization={Springer}
+  organization={Springer},
+  doi={10.1007/978-3-642-17310-3_2}
 }
 
 @misc{audi:2020,
