+++ paper.bib
@@ -58,3 +58,14 @@
   doi={10.1145/3319619.3322003}
 }
 
+@article{turner,
+  title={Introducing a cross platform open source Cartesian Genetic Programming library},
+  author={Turner, Andrew James and Miller, Julian Francis},
+  journal={Genetic Programming and Evolvable Machines},
+  volume={16},
+  number={1},
+  pages={83--91},
+  year={2015},
+  publisher={Springer},
+  doi={10.1007/s10710-014-9233-1}
+}
