+++ paper.bib
@@ -58,7 +58,7 @@
   doi={10.1145/3319619.3322003}
 }
 
-@article{turner,
+@article{turner:2015,
   title={Introducing a cross platform open source Cartesian Genetic Programming library},
   author={Turner, Andrew James and Miller, Julian Francis},
   journal={Genetic Programming and Evolvable Machines},
