### 2020-06-08 #2290 @darioizzo
@shah314

DOIs are added.

We extended a bit on audi, adding a sentence: "The library `audi` makes computations in this algebra rather straight-forward by defining all the basic arithmetics rules for the new type `audi::gdual<T>` which, essentially, represents a multivariate Taylor polynomial and thus contains all the derivatives of the considered expansion."

We commented explicitly upon the absence of crossover in the current version.

We think a comparison between symbolic regression and quadratic/linear regression would be out of scope, maybe a comparison to other existing codes for symbolic regression would be possbile, but its a bit of a pandora box as most are commercial and the non commercial ones are anyway hardly comparable.
