### 2020-06-08 #2290 @darioizzo
@Ohjeah

* We explained at length the concpet of fitness
* We added a last paragraph in the intro to address your point.
