### 2020-06-08 72051cf
C++ example syntax fixed
