### 2020-06-08 #2290 @Ohjeah
- Could you add the statement of need to the online documentation as well?
- Please expand the installtion guide:
  - What versions of the thirdparty dependencies are required?
  - Did you look into an automated package management solution for the c++ version?
  - Why is the PyPI version deprecated? If it's not coming back, please remove the outdated batch from the README.
  - Please add documentation how to execute the tests (I think it's done while installing, you should make the test command explicit).
