+++ README.md
@@ -26,36 +26,106 @@ Preliminary documentation can be found at http://darioizzo.github.io/dcgp/
 
 A web based version of dCGP can be found here: https://esa.github.io/dcgp-web/ thanks to Mike Heddes!
 
-## dcgpy
-The main functionalities of dCGP are exposed into a Python module called dcgpy which can be installed from the conda package ```dcgp-python```  (OSx, linux and Win), pip (deprecated) or by building the module.
+Installation guide
+==================
 
-## Compiling the source code or using the header only library
-### Dependencies
-Several dependencies are necessary to successfully compile d-CGP
- * Audi, headers only library - (git clone https://github.com/darioizzo/audi.git)
- * Boost, headers only
- 
-After making sure the dependencies above are installed in your system, you may download the latest dCGP version via git:
+C++
+---
 
-```git clone https://github.com/darioizzo/d-CGP.git```
+dCGP is a header-only library which has the following third party dependencies:
 
+* [Boost](http://www.boost.org/), various C++ utilities (>=1.72).
+* [Eigen](http://eigen.tuxfamily.org/index.php?title=Main_Page), linear algebra library (>=3.3.0)
+* [Pagmo](https://github.com/esa/pagmo2), parallel optimization library (>=2.15).
+* [tbb](https://github.com/intel/tbb), lets you easily write parallel C++ programs that take full advantage of multicore performance (>=2020.1).
+* [AuDi](http://darioizzo.github.io/audi/), high order automated differentiation library (>=1.8).
+* [Symengine](https://github.com/symengine/symengine), symbolic manipulation of math expressions (>=0.6).
+
+After making sure the dependencies above are installed and found in your system, you may download
+the latest dCGP code via git:
+
+```bash
+git clone https://github.com/darioizzo/dcgp.git
+```
 and configure your build using CMake. 
-
-```cd d-CGP```
-
-```mkdir build```
-
-```cd build```
-
-```ccmake ../```
-
-To build the python package activate the BUILD_DCGPY option.
 
 When done, type (in your build directory):
 
-```make install```
+```bash
+ make 
+```
+When finished, to run the tests type:
 
-The headers will be installed in the CMAKE_INSTALL_PREFIX/include directory. To check that all went well compile the quick-start example.
+```bash
+make test
+```
+
+If succesfull, you may now install cgp:
+
+```bash
+make install
+```
+
+The headers will be installed in the CMAKE_INSTALL_PREFIX/include directory. 
+
+-----------------------------------------------------------------------
+
+Python
+------
+The main functionalities of dCGP are exposed into a Python module called ``dcgpy`` which
+can be installed from conda (OSx, linux and Win), pip (only linux) or by building the module.
+
+### Installing with conda
+
+``dcgpy`` is available in the `conda <https://conda.io/en/latest/>`__ package manager
+from the `conda-forge <https://conda-forge.org/>`__ channel. A single package is available:
+
+* `dcgp-python <https://anaconda.org/conda-forge/dcccgp-python>`__, which contains the ``dcgpy`` python module.
+
+In order to install ``dcgpy`` via conda, you just need
+to add ``conda-forge`` to the channels:
+
+```
+conda config --add channels conda-forge
+conda install dcgp-python
+```
+
+Please refer to the `conda documentation <https://conda.io/en/latest/>`__ for instructions
+on how to setup and manage your conda installation.
+
+You may test the successfull installation by running the python tests typing:
+
+```
+python -c "from dcgpy import test; test.run_test_suite(); import pygmo; pygmo.mp_island.shutdown_pool(); pygmo.mp_bfe.shutdown_pool()"
+```
+
+### Installing with pip (deprecated)
+
[…]
