+++ _sources/notebooks/moth_data.ipynb.txt
@@ -1,288 +1,0 @@
-{
- "cells": [
-  {
-   "cell_type": "code",
-   "execution_count": 42,
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "# Core imports\n",
-    "import dcgpy\n",
-    "import pygmo as pg\n",
-    "import numpy as np\n",
-    "# Sympy is nice to have for basic symbolic manipulation.\n",
-    "from sympy import init_printing\n",
-    "from sympy.parsing.sympy_parser import *\n",
-    "init_printing()\n",
-    "# Plotting\n",
-    "%matplotlib inline\n",
-    "import matplotlib.pyplot as plt\n",
-    "# Pickling\n",
-    "import pickle as pk"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": 43,
-   "metadata": {},
-   "outputs": [],
-   "source": [
-    "# Load the data\n",
-    "with open(\"data.pk\", \"rb\") as f:\n",
-    "    Xs,Ys  = pk.load(f)"
-   ]
-  },
-  {
-   "cell_type": "code",
-   "execution_count": 44,
-   "metadata": {},
-   "outputs": [
-    {
-     "data": {
[…]
