### 2020-06-09 #2290 @darioizzo
@Ohjeah

 * We added the statement in the background section of the web pages
 * Installation guide expanded:
     * all version required of third parties are added
     * guide on readme is now the same as that in the online docs
     * the exact version of the code available in PyPi (it may come back in the future, but we need to find the resources to maintain also those pip packages)
     * no C++ package manager
* Documentation on how tests can be run is added in the installation pages. (for the conda packages those tests are run alreay by conda when creating the packages, but local installation may indeed need to be tested too)
