+++ include/dcgp/kernel_set.hpp
@@ -151,7 +151,14 @@ public:
         return m_kernels[idx];
     }
 
-    // Serialization support.
+    /// Object serialization
+    /**
+     * This method will save/load \p this into the archive \p ar.
+     *
+     * @param ar target archive.
+     *
+     * @throws unspecified any exception thrown by the serialization of the expression and of primitive types.
+     */
     template <typename Archive>
     void serialize(Archive &ar, unsigned)
     {
