+++ include/dcgp/wrapped_functions.hpp
@@ -53,6 +53,7 @@ using f_enabler = typename std::enable_if<std::is_same<T, double>::value || is_g
 
 template <typename T, f_enabler<T> = 0>
 struct my_diff_func {
+    /// Call operator
     T operator()(const std::vector<T> &in) const
     {
         T retval(in[0]);
@@ -68,6 +69,7 @@ template <typename T>
 inline constexpr auto my_diff = my_diff_func<T>{};
 
 struct print_my_diff_func {
+    /// Call operator
     std::string operator()(const std::vector<std::string> &in) const
     {
         std::string retval(in[0]);
@@ -83,6 +85,7 @@ inline constexpr auto print_my_diff = print_my_diff_func{};
 
 template <typename T, f_enabler<T> = 0>
 struct my_mul_func {
+    /// Call operator
     T operator()(const std::vector<T> &in) const
     {
         T retval(in[0]);
@@ -98,6 +101,7 @@ template <typename T>
 inline constexpr auto my_mul = my_mul_func<T>{};
 
 struct print_my_mul_func {
+    /// Call operator
     std::string operator()(const std::vector<std::string> &in) const
     {
         std::string retval(in[0]);
@@ -113,6 +117,7 @@ inline constexpr auto print_my_mul = print_my_mul_func{};
 
 template <typename T, f_enabler<T> = 0>
 struct my_div_func {
+    /// Call operator
     T operator()(const std::vector<T> &in) const
     {
         T retval(in[0]);
@@ -128,6 +133,7 @@ template <typename T>
 inline constexpr auto my_div = my_div_func<T>{};
 
 struct print_my_div_func {
+    /// Call operator
     std::string operator()(const std::vector<std::string> &in) const
     {
         std::string retval(in[0]);
@@ -144,6 +150,7 @@ inline constexpr auto print_my_div = print_my_div_func{};
 // Protected divide function (double overload):
 template <typename T, typename = void>
 struct my_pdiv_func {
+    /// Call operator
     T operator()(const std::vector<T> &in) const
     {
         T retval(in[0]);
@@ -170,6 +177,7 @@ struct my_pdiv_func {
 // Because the gradients created when using gdual are mathematically invalid.
 template <typename T>
 struct my_pdiv_func<T, std::enable_if_t<is_gdual<T>::value>> {
+    /// Call operator
     T operator()(const std::vector<T> &) const
     {
         throw std::invalid_argument("The protected division is not supported for gdual types.");
@@ -181,6 +189,7 @@ template <typename T>
 inline constexpr auto my_pdiv = my_pdiv_func<T>{};
 
 struct print_my_pdiv_func {
+    /// Call operator
     std::string operator()(const std::vector<std::string> &in) const
     {
         std::string retval(in[0]);
@@ -202,6 +211,7 @@ inline constexpr auto print_my_pdiv = print_my_pdiv_func{};
 // sigmoid function: 1 / (1 + exp(- (a + b + c + d+ .. + ))
 template <typename T, f_enabler<T> = 0>
 struct my_sig_func {
+    /// Call operator
     T operator()(const std::vector<T> &in) const
     {
         T retval(in[0]);
@@ -217,6 +227,7 @@ template <typename T>
 inline constexpr auto my_sig = my_sig_func<T>{};
 
 struct print_my_sig_func {
+    /// Call operator
     std::string operator()(const std::vector<std::string> &in) const
     {
         std::string retval(in[0]);
@@ -233,6 +244,7 @@ inline constexpr auto print_my_sig = print_my_sig_func{};
 // tanh function:
 template <typename T, f_enabler<T> = 0>
 struct my_tanh_func {
+    /// Call operator
[…]
