### 2020-06-13 #2290 @shah314
I have been struggling with installing the packages. I was able to install all dependencies (including audi), but the dcgp code is unable to find audi. To install audi from source, I installed all dependencies (including obake), but it is unable to find obake.

Is it possible to write a shell script with the steps needed to install everything? You may be aware of conda command line options that might also speed up the installation. Thank you.
