### 2020-06-13 #2290 @darioizzo
I cannot reproduce your issue. @shah314 can you post the exact conda command you are issuing? From the environment creation. The following for example:

```
conda create -n dcgpy python=3.7 dcgp-python
```

also maybe we can move this in the dcgp GitHub issues?
