### 2020-06-13 darioizzo/dcgp#73 @shah314
obake not found
I was missing the environment creation in the conda command. Now I am able to install everything and run the python tests referenced on the repository. Thank you very much for the help. I think you should change the installation command in the README to the one that you mentioned (that includes the environment creation). I am not very familiar with conda and there could be readers who are also not experts.

But I am still unable to build the git package. The error is as follows:

<pre>
CMake Error at /Users/shah/Documents/anaconda/anaconda3/envs/dcgpy/lib/cmake/audi/audi-config.cmake:9 (find_package):
  By not providing "Findobake.cmake" in CMAKE_MODULE_PATH this project has
  asked CMake to find a package configuration file provided by "obake", but
  CMake did not find one.

  Could not find a package configuration file provided by "obake" with any of
  the following names:

    obakeConfig.cmake
    obake-config.cmake

  Add the installation prefix of "obake" to CMAKE_PREFIX_PATH or set
  "obake_DIR" to a directory containing one of the above files.  If "obake"
  provides a separate development package or SDK, be sure it has been
  installed.
Call Stack (most recent call first):
  CMakeLists.txt:193 (find_package)
</pre>
