### 2020-06-14 darioizzo/dcgp#73 @bluescarni
Can you try to invoke CMake passing the ``CMAKE_PREFIX_PATH`` variable? This needs to point to the path of your conda environment. E.g.,

```
cmake ../ -DCMAKE_PREFIX_PATH=/Users/shah/Documents/anaconda/anaconda3/envs/dcgpy
```
