### 2020-06-14 darioizzo/dcgp#73 @shah314
That gives me the same error. I tried installing obake again but it says it is already installed.

<pre>
(dcgpy) $ sudo conda install -n dcgpy obake
Password:
Collecting package metadata (current_repodata.json): done
Solving environment: done

# All requested packages already installed.

(dcgpy) $
</pre>
