### 2020-06-14 darioizzo/dcgp#73 @bluescarni
obake is split into two packages, ``obake`` (the runtime library) and ``obake-devel`` (headers, CMake files, etc.). If you want to compile something that uses obake, you also need the ``obake-devel`` package installed.
