### 2020-06-14 darioizzo/dcgp#73 @shah314
I see, that fixed the error. Another error:

<pre>
CMake Error at /Users/shah/Documents/anaconda/anaconda3/envs/dcgpy/lib/cmake/Boost-1.72.0/BoostConfig.cmake:185 (boost_find_component):
  boost_find_component Macro invoked with incorrect arguments for macro
  named: boost_find_component
Call Stack (most recent call first):
  /usr/local/Cellar/cmake/3.17.3/share/cmake/Modules/FindBoost.cmake:444 (find_package)
  cmake_modules/DCGPFindBoost.cmake:5 (find_package)
  CMakeLists.txt:183 (include)
</pre>
