### 2020-06-14 darioizzo/dcgp#73 @bluescarni
This is a bug with Boost's CMake files. Try adding the ``-DBoost_NO_BOOST_CMAKE=ON`` option when invoking CMake.
