### 2020-06-18 #2290 @darioizzo
@Ohjeah @shah314 I realize community guidelines are not very visible at the moment, should I add a specific section in the readme file in the github repository?
