### 2020-06-18 d97304f
Update getting_started.py
