### 2020-06-18 #2290 @Ohjeah
Yes you should mention (or link) them in the readme file. Maybe you can have a look into [issue templates](https://help.github.com/en/github/building-a-strong-community/configuring-issue-templates-for-your-repository) as well.
