### 2020-06-19 b1eedc3
fix for cmake instructions
