+++ _sources/quickstart.rst.txt
@@ -22,7 +22,7 @@ To compile it, create also, in the same folder, a ``CmakeLists.txt`` file with t
 
     project(getting_started)
 
-    cmake_minimum_required(VERSION 3.2)
+    cmake_minimum_required(VERSION 3.8)
 
     find_package(dcgp REQUIRED)
 
