### 2020-06-21 #2290 @shah314
@darioizzo There should be a CONTRIBUTING.md and a CODE_OF_CONDUCT.md files present in the repository.

The CODE_OF_CONDUCT can be adapted from [this page](https://www.contributor-covenant.org/version/1/4/code-of-conduct/).

CONTRIBUTING.md could be anything that you want contributors to know before they contribute to the code or submit issues. [Here](https://github.com/shah314/gamultiknapsack/blob/master/CONTRIBUTING.md) is an example.
