### 2020-06-21 #2290 @Ohjeah
I have two remaining points which I would like to see addressed:

- [ ] I'd like to see a more in depth installation guide for the c++ part on how to install the thirdparty requirements. I had to dig into the issues (https://github.com/darioizzo/dcgp/issues/73) and install an undocumented requirement. Maybe the authors could also provide a convenience script to setup a working environment (on macos it's basically a brew + conda install).
- [ ] Similarly the python dependencies should be listed in a standard format (e.g. requirements.txt, environtment.yml, Pipfile or similar). Optional dependencies, as well as dependencies for examples and documentation need to be listed as well. As a user, I'd want to make sure that thirdparty dependencies of dgcpy are compatible before using it in a project.
