### 2020-06-21 #2290 @shah314
@VivianePons @darioizzo @Ohjeah

The code implements differentiable cartesian genetic programming in which the code can learn using symbolic regression. The implementation provides several kernels with the ability to write custom kernels. I tried to fit a few functions from data and it was able to find good approximations.

Another application is the code's ability to learn neural networks and the library provides several activation functions like sigmoid, tanh and ReLu. The python examples provide python notebooks which can be experimented with. In addition to SGD, the library can evolve a neural network's topology using mutation. Comparing this work with [2] might yield some fundamental results.

The work is based on the authors' 2017 paper [1] which has interesting results. The examples in the repository compare the work with a commonly used ANN framework keras and find good results.

Because of several dependencies and interfaces in C++ and python, the initial installation was a minor struggle, but the authors responded to questions fast and resolved the issues.

The only two minor issues I found (in symbolic regression runs) were that, one, the differences between multiple runs were significant. And two, for X and Y of length 10,000, the run was a little slow. Both of these are expected as this is a randomized search procedure with a fixed number of generations.

After reviewing the installation, dependencies and trying out examples, I strongly recommend the work for acceptance to JOSS.

[1] Izzo, Dario, Francesco Biscani, and Alessio Mereta. "Differentiable genetic programming." European Conference on Genetic Programming. Springer, Cham, 2017.
[2] Gaier, Adam, and David Ha. "Weight agnostic neural networks." Advances in Neural Information Processing Systems. 2019.
