### 2020-06-24 darioizzo/dcgp#73 @darioizzo
instructions on how to use conda to install third party dependencies now on the readme.
