+++ include/dcgp/wrapped_functions_s11n_implement.hpp
@@ -37,5 +37,9 @@ DCGP_S11N_FUNCTION_IMPLEMENT_MULTI(my_exp)
 DCGP_S11N_FUNCTION_IMPLEMENT_STRING(print_my_exp)
 DCGP_S11N_FUNCTION_IMPLEMENT_MULTI(my_gaussian)
 DCGP_S11N_FUNCTION_IMPLEMENT_STRING(print_my_gaussian)
+DCGP_S11N_FUNCTION_IMPLEMENT_MULTI(my_sin_nu)
+DCGP_S11N_FUNCTION_IMPLEMENT_STRING(print_my_sin_nu)
+DCGP_S11N_FUNCTION_IMPLEMENT_MULTI(my_cos_nu)
+DCGP_S11N_FUNCTION_IMPLEMENT_STRING(print_my_cos_nu)
 
 #endif
