+++ include/dcgp/kernel_set.hpp
@@ -94,6 +94,8 @@ public:
             m_kernels.emplace_back(my_sin_nu<T>, print_my_sin_nu, kernel_name);
         else if (kernel_name == "cos_nu")
             m_kernels.emplace_back(my_cos_nu<T>, print_my_cos_nu, kernel_name);
+        else if (kernel_name == "gaussian_nu")
+            m_kernels.emplace_back(my_gaussian_nu<T>, print_my_gaussian_nu, kernel_name);
         else
             throw std::invalid_argument("Unimplemented function " + kernel_name + " for this type");
     }
