+++ include/dcgp/wrapped_functions.hpp
@@ -455,7 +455,7 @@ struct print_my_sin_nu_func {
 
 inline constexpr auto print_my_sin_nu = print_my_sin_nu_func{};
 
-// c0sine function (non unary version):
+// cosine function (non unary version):
 template <typename T, f_enabler<T> = 0>
 struct my_cos_nu_func {
     T operator()(const std::vector<T> &in) const
@@ -485,6 +485,37 @@ struct print_my_cos_nu_func {
 };
 
 inline constexpr auto print_my_cos_nu = print_my_cos_nu_func{};
+
+// cosine function (non unary version):
+template <typename T, f_enabler<T> = 0>
+struct my_gaussian_nu_func {
+    T operator()(const std::vector<T> &in) const
+    {
+        T retval(in[0]);
+        for (auto i = 1u; i < in.size(); ++i) {
+            retval += in[i];
+        }
+        return audi::exp(-retval*retval);
+    }
+    DCGP_S11N_EMPTY_SERIALIZE_MEMFN()
+};
+
+template <typename T>
+inline constexpr auto my_gaussian_nu = my_gaussian_nu_func<T>{};
+
+struct print_my_gaussian_nu_func {
+    std::string operator()(const std::vector<std::string> &in) const
+    {
+        std::string retval(in[0]);
+        for (auto i = 1u; i < in.size(); ++i) {
+            retval += "+" + in[i];
+        }
+        return "exp(-(" + retval + ")**2)";
+    }
+    DCGP_S11N_EMPTY_SERIALIZE_MEMFN()
+};
+
+inline constexpr auto print_my_gaussian_nu = print_my_gaussian_nu_func{};
 
 /*--------------------------------------------------------------------------
  *                               UNARY FUNCTIONS
@@ -690,5 +721,11 @@ DCGP_S11N_FUNCTION_EXPORT_KEY_MULTI(my_exp)
 DCGP_S11N_FUNCTION_EXPORT_KEY_STRING(print_my_exp)
 DCGP_S11N_FUNCTION_EXPORT_KEY_MULTI(my_gaussian)
 DCGP_S11N_FUNCTION_EXPORT_KEY_STRING(print_my_gaussian)
+DCGP_S11N_FUNCTION_EXPORT_KEY_MULTI(my_sin_nu)
+DCGP_S11N_FUNCTION_EXPORT_KEY_STRING(print_my_sin_nu)
+DCGP_S11N_FUNCTION_EXPORT_KEY_MULTI(my_cos_nu)
+DCGP_S11N_FUNCTION_EXPORT_KEY_STRING(print_my_cos_nu)
+DCGP_S11N_FUNCTION_EXPORT_KEY_MULTI(my_gaussian_nu)
+DCGP_S11N_FUNCTION_EXPORT_KEY_STRING(print_my_gaussian_nu)
 
 #endif // DCGP_WRAPPED_FUNCTIONS_H
