+++ include/dcgp/wrapped_functions_s11n_implement.hpp
@@ -41,5 +41,7 @@ DCGP_S11N_FUNCTION_IMPLEMENT_MULTI(my_sin_nu)
 DCGP_S11N_FUNCTION_IMPLEMENT_STRING(print_my_sin_nu)
 DCGP_S11N_FUNCTION_IMPLEMENT_MULTI(my_cos_nu)
 DCGP_S11N_FUNCTION_IMPLEMENT_STRING(print_my_cos_nu)
+DCGP_S11N_FUNCTION_IMPLEMENT_MULTI(my_gaussian_nu)
+DCGP_S11N_FUNCTION_IMPLEMENT_STRING(print_my_gaussian_nu)
 
 #endif
