### 2020-07-09 de82bb1
abs step and inv lin
