+++ include/dcgp/expression_ann.hpp
@@ -64,7 +64,13 @@ public:
         /// non unary cosine
         COS_NU,
         /// non unary cosine
-        GAUSSIAN_NU
+        GAUSSIAN_NU,
+        /// negative of the input sum
+        INV_SUM,
+        /// absolute value of inputs
+        ABS,
+        /// step funxtion
+        STEP
     };
     /// Constructor
     /** Constructs a dCGPANN expression
@@ -94,10 +100,11 @@ public:
         for (const auto &ker : f) {
             if (ker.get_name() != "tanh" && ker.get_name() != "sig" && ker.get_name() != "ReLu"
                 && ker.get_name() != "ELU" && ker.get_name() != "ISRU" && ker.get_name() != "sin_nu"
-                && ker.get_name() != "cos_nu" && ker.get_name() != "gaussian_nu" && ker.get_name() != "sum") {
-                throw std::invalid_argument(
-                    "Only tanh, sig, ReLu, ELU, ISRU, sin_nu, cos_nu, gaussian_nu and sum Kernels are valid "
-                    "for dCGP-ANN expressions");
+                && ker.get_name() != "cos_nu" && ker.get_name() != "gaussian_nu" && ker.get_name() != "sum"
+                && ker.get_name() != "inv_sum" && ker.get_name() != "abs" && ker.get_name() != "step") {
+                throw std::invalid_argument("Only tanh, sig, ReLu, ELU, ISRU, sin_nu, cos_nu, gaussian_nu, abs, step "
+                                            "and sum, inv_sum Kernels are valid "
+                                            "for dCGP-ANN expressions");
             }
         }
         // Initialize the kernel map
@@ -120,6 +127,12 @@ public:
                 m_kernel_map[i] = kernel_type::COS_NU;
             } else if (f[i].get_name() == "gaussian_nu") {
                 m_kernel_map[i] = kernel_type::GAUSSIAN_NU;
+            } else if (f[i].get_name() == "inv_sum") {
+                m_kernel_map[i] = kernel_type::INV_SUM;
+            } else if (f[i].get_name() == "abs") {
+                m_kernel_map[i] = kernel_type::ABS;
+            } else if (f[i].get_name() == "step") {
+                m_kernel_map[i] = kernel_type::STEP;
             }
         }
         // Default initialization of weights to 1.
@@ -169,10 +182,11 @@ public:
         for (const auto &ker : f) {
             if (ker.get_name() != "tanh" && ker.get_name() != "sig" && ker.get_name() != "ReLu"
                 && ker.get_name() != "ELU" && ker.get_name() != "ISRU" && ker.get_name() != "sin_nu"
-                && ker.get_name() != "cos_nu" && ker.get_name() != "gaussian_nu" && ker.get_name() != "sum") {
-                throw std::invalid_argument(
-                    "Only tanh, sig, ReLu, ELU, ISRU, sin_nu, cos_nu, gaussian_nu and sum Kernels are valid "
-                    "for dCGP-ANN expressions");
+                && ker.get_name() != "cos_nu" && ker.get_name() != "gaussian_nu" && ker.get_name() != "sum"
+                && ker.get_name() != "inv_sum" && ker.get_name() != "abs" && ker.get_name() != "step") {
+                throw std::invalid_argument("Only tanh, sig, ReLu, ELU, ISRU, sin_nu, cos_nu, gaussian_nu, abs, step "
+                                            "and sum, inv_sum Kernels are valid "
+                                            "for dCGP-ANN expressions");
             }
         }
         // Initialize the kernel map
@@ -195,6 +209,12 @@ public:
                 m_kernel_map[i] = kernel_type::COS_NU;
             } else if (f[i].get_name() == "gaussian_nu") {
                 m_kernel_map[i] = kernel_type::GAUSSIAN_NU;
[…]
