+++ test/expression_ann.cpp
@@ -21,7 +21,7 @@ void test_against_numerical_derivatives(unsigned n, unsigned m, unsigned r, unsi
     std::normal_distribution<> norm{0., 1.};
     std::uniform_int_distribution<unsigned> random_seed(2, 1654636360u);
     // Kernel functions
-    kernel_set<double> ann_set({"sig", "gaussian_nu", "cos_nu", "sin_nu", "ISRU", "sum"});
+    kernel_set<double> ann_set({"sig", "abs", "step", "inv_sum", "ISRU", "sum"});
     // a random dCGPANN
     expression_ann ex(n, m, r, c, lb, arity, ann_set(), random_seed(gen));
     // Since weights and biases are, by default, set to ones, we randomize them
