+++ include/dcgp/wrapped_functions_s11n_implement.hpp
@@ -43,5 +43,11 @@ DCGP_S11N_FUNCTION_IMPLEMENT_MULTI(my_cos_nu)
 DCGP_S11N_FUNCTION_IMPLEMENT_STRING(print_my_cos_nu)
 DCGP_S11N_FUNCTION_IMPLEMENT_MULTI(my_gaussian_nu)
 DCGP_S11N_FUNCTION_IMPLEMENT_STRING(print_my_gaussian_nu)
+DCGP_S11N_FUNCTION_IMPLEMENT_MULTI(my_inv_sum)
+DCGP_S11N_FUNCTION_IMPLEMENT_STRING(print_my_inv_sum)
+DCGP_S11N_FUNCTION_IMPLEMENT_MULTI(my_abs)
+DCGP_S11N_FUNCTION_IMPLEMENT_STRING(print_my_abs)
+DCGP_S11N_FUNCTION_IMPLEMENT_MULTI(my_step)
+DCGP_S11N_FUNCTION_IMPLEMENT_STRING(print_my_step)
 
 #endif
