### 2020-07-10 aee34d0
docs update for code review
