+++ doc/sphinx/docs/cpp/kernel_list.rst
@@ -53,3 +53,15 @@ kernels to the set.
    +----------------+------------------------+---------------------------------------------------------------------------------------+
    |"sum"           |addition                | :math:`\sum_i x_i`                                                                    |
    +----------------+------------------------+---------------------------------------------------------------------------------------+
+   |"sin_nu"        |sine (non unary)        | :math:`\sin(\sum_i x_i)`                                                              |
+   +----------------+------------------------+---------------------------------------------------------------------------------------+
+   |"cos_nu"        |cosine (non unary)      | :math:`\cos(\sum_i x_i)`                                                              |
+   +----------------+------------------------+---------------------------------------------------------------------------------------+
+   |"gaussian_nu"   |gaussian (non unary)    | :math:`e^{\sum_i x_i}`                                                                |
+   +----------------+------------------------+---------------------------------------------------------------------------------------+
+   |"abs"           |absolute value          | :math:`\abs(\sum_i x_i)`                                                              |
+   +----------------+------------------------+---------------------------------------------------------------------------------------+
+   |"inv_sum"       |inverse sum             | :math:`- \sum_i x_i`                                                                  |
+   +----------------+------------------------+---------------------------------------------------------------------------------------+
+   |"step"          |step function           | 1 if positive, 0 otherwise                                                            |
+   +----------------+------------------------+---------------------------------------------------------------------------------------+
