+++ doc/sphinx/installation.rst
@@ -13,7 +13,14 @@ dCGP is a header-only library which has the following third party dependencies:
 * `Pagmo <https://github.com/esa/pagmo2>`_, parallel optimization library. (>=2.15).
 * `tbb <https://github.com/intel/tbb>`_, lets you easily write parallel C++ programs that take full advantage of multicore performance. (>=2020.1).
 * `AuDi <http://darioizzo.github.io/audi/>`_, high order automated differentiation library. (>=1.8).
+* `obake <https://github.com/bluescarni/obake>`,  symbolic manipulation of sparse polynomials. (>=0.6.0). 
 * `Symengine <https://github.com/symengine/symengine>`_, symbolic manipulation of math expressions. (>=0.6).
+
+In case you are familiar with the conda package manager an environent ready for the dcgp installation can be created via the single command
+
+.. code-block:: console
+
+   $ conda create -n build_dcgp cmake cxx-compiler boost-cpp pagmo-devel tbb-devel audi symengine obake-devel
 
 After making sure the dependencies above are installed and found in your system, you may download
 the latest dCGP code via git:
@@ -22,7 +29,13 @@ the latest dCGP code via git:
 
    $ git clone https://github.com/darioizzo/dcgp.git
 
-and configure your build using CMake. 
+and configure your build using CMake. For example, if you are using conda, make a build directory and type:
+
+.. code-block:: console
+
+   $ cmake -DBoost_NO_BOOST_CMAKE=ON -DCMAKE_INSTALL_PREFIX=~/miniconda3/envs/build_dcgp/ -DCMAKE_PREFIX_PATH=~/miniconda3/envs/build_dcgp/ ../
+
+where the conda environment has been assumed in ```~/miniconda3/envs/build_dcgp/```.
 
 When done, type (in your build directory):
 
@@ -78,6 +91,8 @@ to add ``conda-forge`` to the channels:
    $ conda config --add channels conda-forge
    $ conda install dcgp-python
 
+note that all the required dependencies will be installed automatically, as well as the C++ ```dcgp``` headers.
+
 Please refer to the `conda documentation <https://conda.io/en/latest/>`__ for instructions
 on how to setup and manage your conda installation.
 
@@ -88,24 +103,33 @@ You may test the successfull installation by running the python tests typing:
    $ python -c "from dcgpy import test; test.run_test_suite(); import pygmo; pygmo.mp_island.shutdown_pool(); pygmo.mp_bfe.shutdown_pool()"
 
 
-Installing with pip (deprecated)
-^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-We also provide the pip packages (mainly for linux 64 bit architectures and versions <= 1.4.1).
-Check on the `PyPi dcgpy page <https://pypi.org/project/dcgpy/>`_ if the needed package is provided.
-
-.. code-block:: console
-
-   $ pip install dcgpy
-
 Building
 ^^^^^^^^^^^^^^^^^^^^^^^^^^
 
-To build the module you need to have the Boost Python libraries installed and to activate the BUILD_DCGPY option from within CMake (and deselect BUILD_DCGP)
+To build the python module you need to first install the dcgp C++ header library and its dependencies (see above) as well as the additional dependency:
 
-Check carefully what Python version is detected and what libraries are linked to. In particular select the correct boost_python
-according to the Python version (2 or 3) you want to compile the module for.
+* `pybind11 <https://github.com/pybind/pybind11>`, Seamless operability between C++11 and Python. (>=2.5.0). 
 
-The CMAKE_INSTALL_PREFIX will be used to construct the final location of headers and Python module after install.
[…]
