+++ doc/sphinx/docs/cpp/kernel_list.rst
@@ -59,7 +59,7 @@ kernels to the set.
    +----------------+------------------------+---------------------------------------------------------------------------------------+
    |"gaussian_nu"   |gaussian (non unary)    | :math:`e^{\sum_i x_i}`                                                                |
    +----------------+------------------------+---------------------------------------------------------------------------------------+
-   |"abs"           |absolute value          | :math:`\abs(\sum_i x_i)`                                                              |
+   |"abs"           |absolute value          | :math:`\vert\sum_i x_i\vert`                                                          |
    +----------------+------------------------+---------------------------------------------------------------------------------------+
    |"inv_sum"       |inverse sum             | :math:`- \sum_i x_i`                                                                  |
    +----------------+------------------------+---------------------------------------------------------------------------------------+
