### 2020-07-10 #2290 @darioizzo
@Ohjeah Sorry for the inactivity. I have implemented your requested changes in the documentation: http://darioizzo.github.io/dcgp/installation.html

I added the exact commands, assuming a conda user toset up correctly all dependencies.
