### 2020-07-10 ab45683
bug fix for retun of second obj
