+++ test/symbolic_regression.cpp
@@ -118,7 +118,14 @@ BOOST_AUTO_TEST_CASE(fitness_test_two_obj)
                   - static_cast<decltype(string.length())>(std::count(string.begin(), string.end(), ' '))
                   - 2u; // no spaces and no [] parenthesis
         auto l2 = udp.pretty(pop.get_x()[i]).length() - 2u;
-        BOOST_CHECK_EQUAL(std::min(l1, l2), pop.get_f()[i][1]);
+        if (std::isfinite(pop.get_f()[i][0])) {
+            BOOST_CHECK_EQUAL(std::min(l1, l2), pop.get_f()[i][1]);
+        } else {
+            BOOST_CHECK_EQUAL(l2, pop.get_f()[i][1]);
+        }
+std::cout << "\npretty: " << udp.pretty(pop.get_x()[i]) << "\n";
+std::cout << "prettier: " << string << "\n";
+std::cout << "l1: " << l1 << "\tl2: " << l2 << "\tfit: " << pop.get_f()[i][1] << "\n";
     }
 }
 
