### 2020-07-10 95e020b
debug streams deleted
