### 2020-07-13 #2290 @Ohjeah
@darioizzo Thanks! Have you looked into [environment files](https://docs.conda.io/projects/conda/en/latest/user-guide/tasks/manage-environments.html#create-env-file-manually).

I am only missing the Community guidelines. A good example would be using Github issue templates and a paragraph in the Readme similar to this project: https://github.com/BartoszBartmanski/StoSpa2
