### 2020-07-15 #2290 @darioizzo
@Ohjeah Thanks for your suggestions, I have added issue templates in github and community guidelines in the readme.
