### 2020-07-15 #2290 @darioizzo
@VivianePons we have this zenodo record already: https://zenodo.org/record/3802555#.Xw8ETygzaUk, it is linked to all new releases ...
