### 2020-07-15 #2290 @VivianePons
Congratulations! I can now accept the paper. This is a very nice work and I am happy to see it published at JOSS. Thank you very much @darioizzo for submitting it and thanks to @shah314 and @Ohjeah for the reviews and feedbacks
