### 2020-07-15 #2290 @labarba
@darioizzo -- version tag 1.5 in the repo is from May 6. Please confirm: there have been no changes since due to this review? (if yes, you could up the patch number in a new tag)
