### 2020-07-16 #2290 @darioizzo
@labarba 1.5 is the last released code and is what the paper refers to. There have been minor changes in the repo since we are developing 1.5.1, but it is not ready for a release yet, that is planned in 6-12 months.
