### 2020-07-16 #2290 @darioizzo
@labarba The branch was created as per request during the review process (orignally I had put the paper into an entirely different repo), I gave it the name 1.5.1 back then as I took it from the development code (in retrospective I should have used 1.5).

There is not any difference (only code overhauls) in features between 1.5 and 1.5.1. But it is true that 1.5.1 is under development and unreleased.
