### 2020-07-16 #2290 @darioizzo
I can clone into a new branch with the 1.5 name ,if you want, and delete the old branch?
