### 2020-07-16 #2290 @labarba
Congratulation, @darioizzo, your JOSS paper is published! 🚀

Huge thanks to our editor: @VivianePons, and the reviewer: @Ohjeah, @shah314 🙏
